Fix default foreground and background colours
It turns out that reading `tooltip' face parameters when providing
default values for defcustom-defined variables does not always work
correctly. In particular it causes certain problems when Emacs is
started in daemon mode, i.e., it doesn't have a graphical frame at
first, and then one is created.

This moves the behaviour of using `tooltip' face by default to point of
use of `pos-tip-foreground-color', and `pos-tip-background-color',
adding a new possible value for this, `nil', that will cause runtime
lookup of `tooltip' face properties.