Fix face attributes applied by `pos-tip-show'
The font parameter of a frame is a string, which is not a valid font
object, obviously. It can still be used as :font parameter of a face in
certain circumstances, but `propertize' calls seem not to be one of
those, leading to warnings about invalid face attributes.

So now `pos-tip-show' will actually construct a font-spec out of it, and
find the corresponding font object.

Fixes #10.